ANDDEP-82 Внесены мелкие правки по коду семпла. Доработаны Security tips

+++ network-sample/src/main/java/ru/surfstudio/android/network/sample/interactor/common/network/OkHttpModule.kt
@@ -13,7 +13,7 @@ import java.util.concurrent.TimeUnit
 class OkHttpModule {
 
     companion object {
-        private const val NETWORK_TIMEOUT = 10 //sec
+        private const val NETWORK_TIMEOUT = 10L //sec
     }
 
     @Provides
@@ -22,9 +22,9 @@ class OkHttpModule {
                             etagInterceptor: EtagInterceptor,
                             httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
         return OkHttpClient.Builder().apply {
-            connectTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
-            readTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
-            writeTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
+            connectTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
+            readTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
+            writeTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
 
             addInterceptor(cacheInterceptor)
             addInterceptor(etagInterceptor)
